feat: Add custom_user_plugin to DNS server and frontend
This commit adds the custom_user_plugin to the DNS server and frontend. The plugin is registered in the dnsserver package and imported in the zdirectives.go file. In the zplugin.go file, the plugin is imported and added to the list of plugins. Additionally, the necessary configuration files (.npmrc and README.md) are added to the dns-manager-frontend directory. Finally, the setup.go file for the custom_user_plugin is added to the plugin/custom_user_plugin directory.